# icp-handler/openapi.yaml
openapi: 3.1.0
info:
  title: ICP-1.0 Merchant Handler API
  version: 1.0.0
  description: |
    Normative HTTP API exposed by an ICP-1.0 merchant handler. Conformant
    handlers MUST implement every operation in this document with the
    request/response shapes described here.

    This spec is generated from `icp-spec/ICP-1.0-DRAFT.md` and the JSON
    Schemas in `icp-spec/schemas/`. It is intended to drive client codegen
    in any language without a hand-rolled SDK (Java / C# / Swift / Kotlin
    / Ruby / PHP / Dart / Elixir / etc).

    **Canonicalization:** every Intent submitted to the handler MUST be
    serialized using RFC 8785 Canonical JSON before signing. The handler
    re-canonicalizes the parsed body and verifies the signature against
    the result. Any deviation in canonical bytes produces
    `signature.invalid` (HTTP 401). See
    `icp-spec/schemas/canonicalization.md` for full rules.

    **Error model:** all error responses share the `Error` envelope and
    use ICP error codes from `icp-spec/schemas/error-codes.md`. HTTP
    status mappings are normative.

  contact:
    name: StateSet (reference implementer)
    url: https://github.com/stateset/stateset-icommerce
  license:
    name: Apache-2.0 (code) / CC-BY-4.0 (spec)
    url: https://github.com/stateset/stateset-icommerce/blob/master/LICENSE
servers:
  - url: https://shop.example.com
    description: Production handler (over TLS)
  - url: http://127.0.0.1:7402
...
paths:
...
  /icp/v1/escrows/{escrow_id}/events:
    get:
      tags: [escrow]
      summary: Subscribe to escrow event stream (SSE)
      description: |
        Server-Sent Events. Emits one event per escrow state transition.
        Event types: `escrow.opened`, `escrow.funded`, `escrow.fulfilled`,
        `escrow.released`, `escrow.refunded`, `dispute.opened`,
        `dispute.resolved`. The merchant signs every event payload; the
        caller MUST verify against the merchant's well-known pubkey.

        Heartbeats are sent every 30s as `:heartbeat` comment lines so
        idle proxies don't close the stream.
      operationId: observe
      parameters:
        - $ref: '#/components/parameters/EscrowId'
      responses:
        '200':
          description: SSE stream of escrow events
          content:
            text/event-stream:
              schema:
                type: string
                description: Newline-delimited SSE frames
              example: |
                event: escrow.opened
                data: {"escrow_id":"0xabc...","amount":{"amount":"29.99","currency":"USDC"},"signature_hex":"…"}

                event: escrow.funded
                data: {"escrow_id":"0xabc...","tx_hash":"0xfee…","signature_hex":"…"}

        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
...
    EscrowId:
      name: escrow_id
      in: path
      required: true
      description: Escrow identifier returned by `acceptQuote`
      schema: { $ref: '#/components/schemas/EscrowId' }
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            type: icp.error
            code: format.not_found
            message: "no such quote"
...
  schemas:
...
    EscrowId:
      type: string
      description: Opaque Settler-issued identifier (often a 0x… hash on EVM rails)
      examples: ['0x3b8c…']
...
    Error:
      type: object
      required: [type, code, message]
      description: |
        Uniform error envelope. `code` is from the ICP error code
        registry in `icp-spec/schemas/error-codes.md`.
      properties:
        type:
          type: string
          const: icp.error
        code:
          type: string
          description: Dotted-namespace code. See `error-codes.md`.
          examples:
            - signature.invalid
            - policy.settler.not_allowed
            - replay.duplicate_nonce
            - escrow.invalid_state
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...